added get delivary to context, changed decimal to int on all modals, added locations.html to hold the delivary form

+++ checkout/models.py
@@ -27,12 +27,9 @@ class Order(models.Model):
     street_address2 = models.CharField(max_length=80, null=True, blank=True)
     county = models.CharField(max_length=80, null=True, blank=True)
     date = models.DateTimeField(auto_now_add=True)
-    delivery_cost = models.DecimalField(
-        max_digits=6, decimal_places=2, null=False, default=0)
-    order_total = models.DecimalField(
-        max_digits=10, decimal_places=2, null=False, default=0)
-    grand_total = models.DecimalField(
-        max_digits=10, decimal_places=2, null=False, default=0)
+    delivery_cost = models.IntegerField( null=False, default=0)
+    order_total = models.IntegerField( null=False, default=0)
+    grand_total = models.IntegerField( null=False, default=0)
     original_bag = models.TextField(null=False, blank=False, default='')
     stripe_pid = models.CharField(
         max_length=254, null=False, blank=False, default='')
@@ -88,8 +85,7 @@ class OrderLineItem(models.Model):
     product_size = models.CharField(
         max_length=2, null=True, blank=True)
     quantity = models.IntegerField(null=False, blank=False, default=0)
-    lineitem_total = models.DecimalField(max_digits=6, decimal_places=2, 
-                                         null=False, blank=False, editable=False)
+    lineitem_total = models.IntegerField(null=False, blank=False, editable=False)
     
     def save(self, *args, **kwargs):
         """
